Enable re-login
This PR makes the client persist correctly in the case that you're running it in a loop. Unfortunately, if you have a policy that requires MFA, you will have to re-authenticate with the MFA each time it refreshes, but this is the best we can do for now.

see this for more context from the onelogin team: https://stackoverflow.com/questions/40336144/aws-api-credentials-with-onelogin-saml-and-mfa/47065404#47065404

* Factor out OTP validation into a separate method
* Enable retries on login and OTP validation to enable mistakes
* Reuse `user_choice` method for getting factors