Get all prompt to use user_choice
In preparation for being able to save prompt selections, this moves more of the prompts to use user_choice.  The function is extended to allow custom renderers, and adds the behavior we have in most prompts (e.g., automatic selection if only one option is available); this way we can kill a bunch of redundant code.

This also simplifies the way we store the current device somewhat.